requirements.txt
streamlit
google-generativeai
langchain
PyPDF2
chromadb
python-dotenv